requirements.txt
mcp[cli]==1.26.0
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
